% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancorr.R
\name{print.CanCorr}
\alias{print.CanCorr}
\title{Print Method for CanCorr Objects}
\usage{
\method{print}{CanCorr}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object returned from the class of \code{CanCorr}.}

\item{digits}{Number of digits in printing the matrices. The default is 4.}

\item{...}{Unused.}
}
\value{
No return value, called for side effects
}
\description{
Print method for \code{CanCorr} objects.
}
\author{
Mike W.-L. Cheung \href{mailto:mikewlcheung@nus.edu.sg}{mikewlcheung@nus.edu.sg}
}
