% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcr.R
\name{print.MPCR}
\alias{print.MPCR}
\title{Print Method for MPCR Objects}
\usage{
\method{print}{MPCR}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object returned from the class of \code{MPCR}.}

\item{digits}{Number of digits in printing the matrices. The default is 4.}

\item{...}{Unused.}
}
\value{
No return value, called for side effects
}
\description{
Print method for \code{MPCR} objects.
}
\author{
Mike W.-L. Cheung \href{mailto:mikewlcheung@nus.edu.sg}{mikewlcheung@nus.edu.sg}
}
