\name{print}
\alias{print.RA}
\title{Print Methods for various Objects 
}
\description{Print method for the \code{RA} object.
}
\usage{
\method{print}{RA}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object returned from the class
	\code{RA}
}
  \item{\dots}{Unused.
}
}

\author{Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
