% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rda.R
\name{print.RDA}
\alias{print.RDA}
\title{Print Method for RDA Objects}
\usage{
\method{print}{RDA}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object returned from the class of \code{RDA}.}

\item{digits}{Number of digits in printing the matrices. The default is 4.}

\item{...}{Unused.}
}
\value{
No return value, called for side effects
}
\description{
Print method for \code{RDA} objects.
}
\author{
Mike W.-L. Cheung \href{mailto:mikewlcheung@nus.edu.sg}{mikewlcheung@nus.edu.sg}
}
