% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sas_ex2}
\alias{sas_ex2}
\title{Sample data for redundancy analysis from the SAS manual}
\format{
A matrix with 10 rows and 7 columns:
\describe{
\item{y1, y2, y3}{Y variables}
\item{x1, x2, x3, x4}{X variables}
}
}
\source{
\url{https://documentation.sas.com/doc/en/pgmsascdc/9.4_3.3/statug/statug_transreg_details23.htm}
}
\usage{
sas_ex2
}
\description{
This dataset includes seven variables from the SAS manual.
}
\examples{
data(sas_ex2)

\donttest{
## Redundancy Analysis
rda(X_vars=c("x1", "x2", "x3", "x4"),
    Y_vars=c("y1", "y2", "y3"),
    data=sas_ex2)
}
}
\keyword{datasets}
