% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lambert88}
\alias{Lambert88}
\title{Correlation matrix of artificial data}
\format{
A 9x9 correlation matrix.
}
\source{
Lambert, Z. V., Wildt, A. R., & Durand, R. M. (1988). Redundancy
analysis: An alternative to canonical correlation and multivariate
multiple regression in exploring interset associations.
\emph{Psychological Bulletin}, \strong{104(2)}, 282-289.
\doi{10.1037/0033-2909.104.2.282}
}
\usage{
Lambert88
}
\description{
This dataset includes a correlation matrix of nine artificial
variables used in Table 1 of Lambert, Wildt, and Durand (1988).
}
\examples{
data(Lambert88)

\donttest{
## Redundancy Analysis
rda(X_vars=paste0("x", 1:5), Y_vars=paste0("y", 1:4), Cov=Lambert88, numObs=100)
}
}
\references{
Gu, F., Yung, Y.-F., Cheung, M. W.-L., Joo, B.-K., & Nimon,
K. (2023). Statistical inference in redundancy analysis: A direct
covariance structure modeling approach. \emph{Multivariate Behavioral
Research},
\strong{58(5)}, 877-893. \doi{10.1080/00273171.2022.2141675}
}
\keyword{datasets}
