\name{print}
\alias{print.CanCorr}
\alias{print.RDA}
\title{Print Methods for various Objects 
}
\description{Print method for \code{CanCorr} and \code{RDA} objects.
}
\usage{
\method{print}{CanCorr}(x, \dots)
\method{print}{RDA}(x, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object returned from the class of either \code{CanCorr} or
	\code{RDA}.
}
  \item{\dots}{Unused.
}
}

\author{Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
