\name{print}
\alias{print.CanCorr}
\alias{print.RDA}
\alias{print.MPCR}
\title{Print Methods for various Objects 
}
\description{Print method for \code{CanCorr} and \code{RDA} objects.
}
\usage{
\method{print}{CanCorr}(x, digits=4, \dots)
\method{print}{RDA}(x, digits=4, \dots)
\method{print}{MPCR}(x, digits=4, \dots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{An object returned from the class of either \code{CanCorr},
	\code{RDA}, or \code{MPCR}.
  }
  \item{digits}{Number of digits in printing the matrices. The default
	is 4.}
  \item{\dots}{Unused.}
}

\author{Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>
}
